src/main/java/hello/core/discount/RateDiscountPolicy.java
package hello.core.discount;

import hello.core.annotation.MainDiscountPolicy;
import hello.core.member.Grade;
import hello.core.member.Member;
import org.springframework.beans.factory.annotation.Qualifier;
import org.springframework.context.annotation.Primary;
import org.springframework.stereotype.Component;

@Component
@MainDiscountPolicy
// 할인 정책 구현체 (2. 정률 할인)
//@Qualifier("mainDiscountPolicy") // 같은 타입의 빈이 2개 이상인 경우 => @Component 이전에 이름 매핑 (@Componenet 보다 우선순위가 빠르다)
@Primary
public class RateDiscountPolicy implements DiscountPolicy{

    private int discountPercent = 10;

    @Override
    public int discount(Member member, int price) {
        if (member.getGrade() == Grade.VIP) {
            return price * discountPercent / 100;
        } else {
            return 0;
        }
    }
}
